Finalize auth model, domain, permission etc ... Use test db for unit test. Upgrade angular-translate

+++ modules/web-services/restfull/src/main/java/org/constellation/ws/rest/UserService.java
@@ -28,10 +28,12 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import org.constellation.configuration.AcknowlegementType;
+import org.constellation.engine.register.DomainUser;
 import org.constellation.engine.register.User;
 import org.constellation.engine.register.repository.UserRepository;
 import org.geotoolkit.util.StringUtilities;
@@ -49,80 +51,58 @@ import org.springframework.util.StringUtils;
 @Path("/1/user")
 @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
 @Consumes({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
-public class UserService implements IUserService {
+public class UserService  {
 
 
     @Inject
     private UserRepository userRepository;
     
-   
-    /*
-     * (non-Javadoc)
-     * 
-     * @see org.constellation.ws.rest.IUserService#findOne(java.lang.String)
-     */
-    @Override
+    @GET
+    @Path("/")
+    public Response findAll(@QueryParam("withDomainAndRoles") boolean withDomainAndRole) {
+        if(withDomainAndRole) {
+            return Response.ok(userRepository.findAllWithDomainAndRole()).build();
+        }
+        return Response.ok(userRepository.findAll()).build();
+    }
+
     @GET
     @Path("/{id}")
     public Response findOne(@PathParam("id") String login) {
-        return Response.ok(userRepository.all()).build();
+        return Response.ok(userRepository.findOneWithRolesAndDomains(login)).build();
     }
 
-    /*
-     * (non-Javadoc)
-     * 
-     * @see org.constellation.ws.rest.IUserService#findAll(java.lang.String)
-     */
-    @Override
-    @GET
-    @Path("/")
-    public Response findAll(@PathParam("id") String login) {
-        return Response.ok(userRepository.all()).build();
-    }
 
-    /*
-     * (non-Javadoc)
-     * 
-     * @see org.constellation.ws.rest.IUserService#delete(java.lang.String)
-     */
-    @Override
     @DELETE
     @Path("/{id}")
     public Response delete(@PathParam("id") String id) {
+        if(userRepository.isLastAdmin(id))
+            return Response.serverError().build();
         userRepository.delete(id);
         return Response.noContent().build();
     }
 
-    /*
-     * (non-Javadoc)
-     * 
-     * @see
-     * org.constellation.ws.rest.IUserService#post(org.constellation.engine.
-     * register.UserDTO)
-     */
-    @Override
     @POST
     @Path("/")
     @Transactional
-    public Response post(User userDTO) {
+    public Response post(DomainUser userDTO) {
         if (StringUtils.hasText(userDTO.getPassword()))
             userDTO.setPassword(StringUtilities.MD5encode(userDTO.getPassword()));
 
-        userRepository.insert(userDTO);
+        userRepository.insert(userDTO, userDTO.getRoles());
 
         return Response.ok(userDTO).build();
     }
     
     
-    @Override
     @PUT
     @Path("/")
     @Transactional
-    public Response put(User userDTO) {
+    public Response put(DomainUser userDTO) {
         if (StringUtils.hasText(userDTO.getPassword()))
             userDTO.setPassword(StringUtilities.MD5encode(userDTO.getPassword()));
 
-        userRepository.update(userDTO);
+        userRepository.update(userDTO, userDTO.getRoles());
 
         return Response.ok(userDTO).build();
     }
